+++ packages/backend/src/modules/interop/plugins/hyperlane-hwr.ts
@@ -23,6 +23,7 @@ import {
 import { findHyperlaneChain, HyperlaneConfig } from './hyperlane.config'
 import { getBridgeType } from './layerzero/layerzero-v2-ofts.plugin'
 import { findParsedAround, type ParsedTransferLog } from './logScan'
+import { getBestEffortTokenFrameworkBridgeType } from './tokenFrameworkBridgeTyping'
 import {
   createEventParser,
   createInteropEventType,
@@ -284,7 +285,10 @@ export class HyperlaneHwrPlugin implements InteropPluginResyncable {
             dstTokenAddress: event.args.tokenAddress,
             dstAmount: event.args.amount,
             dstWasMinted: event.args.minted,
-            bridgeType: 'burnAndMint',
+            bridgeType: getBestEffortTokenFrameworkBridgeType({
+              srcWasBurned: undefined,
+              dstWasMinted: event.args.minted,
+            }),
           }),
         ]
       }
@@ -365,7 +369,10 @@ export class HyperlaneHwrPlugin implements InteropPluginResyncable {
         srcAmount: event.args.amount,
         srcWasBurned: event.args.burned,
         dstChain,
-        bridgeType: 'burnAndMint',
+        bridgeType: getBestEffortTokenFrameworkBridgeType({
+          srcWasBurned: event.args.burned,
+          dstWasMinted: undefined,
+        }),
       }),
     ]
   }
